% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eic.R
\name{eic}
\alias{eic}
\title{Get the Energy Identification Codes (EIC)}
\usage{
eic()
}
\description{
Read more about the EIC here: https://www.entsoe.eu/fileadmin/user_upload/edi/library/downloads/EIC_Reference_Manual.pdf
}
\examples{
library(entsoeR)

eic()
}
