% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_parallel.R
\name{fetch_parallel}
\alias{fetch_parallel}
\title{Fetch data from ENTSO-E csv files, using multiple cores (to speed up process)}
\usage{
fetch_parallel(
  dataset = "LOAD",
  areas = "all",
  aggregate.hourly = TRUE,
  from = lubridate::floor_date(Sys.time(), "year"),
  to = lubridate::floor_date(Sys.time(), "month") - 86400,
  technology = NULL,
  units = NULL
)
}
\arguments{
\item{dataset}{is used to select data domain, e.g. "load", "price", or "generation"}

\item{areas}{is a list of shorthand area names to subset data}

\item{aggregate.hourly}{if TRUE returns values by hour, as sometimes data is more granular}

\item{from}{date to start time series}

\item{to}{date to end time series}

\item{technology}{(optional) is used to subset data by generation technology}

\item{units}{(optional) is used to subset data by generation unit}
}
\value{
a data frame containing the selected data from ENTSO-E
}
\description{
A universal function to collect and compile data from
ENTSO-E's transparency platform, downloaded as csv files
via the sftp server.
}
\details{
For dodumentation, see
https://transparency.entsoe.eu/content/static_content/Static\%20content/knowledge\%20base/SFTP-Transparency_Docs.html
https://transparency.entsoe.eu/content/static_content/Static\%20content/web\%20api/Guide.html

This function requires having downloaded data as monthly .csv-files via ENTSO-E sftp server, and stored in folder
}
\examples{
# Examples of how to use the function

\dontrun{
load <- fetch_parallel(dataset = "Generation", areas = c("DK1", "DK2"),
            technologies = c("Wind Offshore", "Wind Onshore"),
            from = "2024-01-01", to = "2024-02-01")
}

}
\references{
\url{https://transparency.entsoe.eu/content/static_content/Static\%20content/web\%20api/Guide.html}
}
\author{
Magnus Hornoe Gottlieb
}
\keyword{ENTSO-E}
\keyword{Energy}
\keyword{Europe}
\keyword{Grid}
\keyword{Power}
