% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generation_get.R
\name{generation_get}
\alias{generation_get}
\title{This wraps a GET request to the API}
\usage{
generation_get(documentType = NULL, processType = NULL,
  businessType = NULL, psrType = NULL, type_MarketAgreement.Type = NULL,
  contract_MarketAgreement.Type = NULL, auction.Type = NULL,
  auction.Category = NULL,
  classificationSequence_AttributeInstanceComponent.Position = NULL,
  outBiddingZone_Domain = NULL, biddingZone_Domain = NULL,
  controlArea_Domain = NULL, in_Domain = NULL, out_Domain = NULL,
  acquiring_Domain = NULL, timeInterval = NULL, periodStart = NULL,
  periodEnd = NULL, securityToken = Sys.getenv("ENTSOE_PAT"))
}
\description{
There's a limit of 200 documents, so the API will return an
error if this limit is reached.
}
\examples{

library(entsoeR)

# 
# https://transparency.entsoe.eu/content/static_content/Static\%20content/web\%20api/Guide.html#_outages
# 4.4.1. Installed Generation Capacity Aggregated [14.1.A]
generation_get(documentType = "A68",
               processType = "A33",
               psrType = "B16",
               in_Domain = "10YCZ-CEPS-----N",
               periodStart = "201512312300",
               periodEnd = "201612312300")
 
# 4.4.2. Installed Generation Capacity per Unit [14.1.B]
generation_get(documentType = "A71",
               processType = "A33",
               psrType = "B02",
               in_Domain = "10YCZ-CEPS-----N",
               periodStart = "201512312300",
               periodEnd = "201612312300")
 
# 4.4.3. Day-ahead Aggregated Generation [14.1.C]
generation_get(documentType = "A71",
               processType = "A01",
               in_Domain = "10YCZ-CEPS-----N",
               periodStart = "201512312300",
               periodEnd = "201612312300")
 
# 4.4.4. Day-ahead Generation Forecasts for Wind and Solar [14.1.D]
generation_get(documentType = "A69",
               processType = "A01",
               psrType = "B16",
               in_Domain = "10YCZ-CEPS-----N",
               periodStart = "201512312300",
               periodEnd = "201612312300")
 
# 4.4.5. Actual Generation Output per Generation Unit [16.1.A]
generation_get(documentType = "A73",
               processType = "A16",
               psrType = "B02",
               in_Domain = "10YCZ-CEPS-----N",
               periodStart = "201512312300",
               periodEnd = "201612312300")
 
# 4.4.5. Actual Generation Output per Generation Unit [16.1.A]
generation_get(documentType = "A73",
               processType = "A16",
               psrType = "B02",
               in_Domain = "10YCZ-CEPS-----N",
               periodStart = "201512312300",
               periodEnd = "201612312300")
 
}
