% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generation_get_ts.R
\name{generation_get_ts}
\alias{generation_get_ts}
\title{time series version of generation_get}
\usage{
generation_get_ts(documentType = NULL, processType = NULL,
  businessType = NULL, psrType = NULL, type_MarketAgreement.Type = NULL,
  contract_MarketAgreement.Type = NULL, auction.Type = NULL,
  auction.Category = NULL,
  classificationSequence_AttributeInstanceComponent.Position = NULL,
  outBiddingZone_Domain = NULL, biddingZone_Domain = NULL,
  controlArea_Domain = NULL, in_Domain = NULL, out_Domain = NULL,
  acquiring_Domain = NULL, timeInterval = NULL, periodStart = NULL,
  periodEnd = NULL, securityToken = NULL)
}
\description{
function to download generation data in a "time series" format (dates, generation)
}
