% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_folder_path.R
\name{get_folder_path}
\alias{get_folder_path}
\title{Get Folder Path for CSV Files}
\usage{
get_folder_path()
}
\value{
A character string representing the folder path.
}
\description{
This function checks if the folder path containing the CSV files is already set.
If the path is not set, it prompts the user to input the path, validates it,
and then saves it for future use.
}
