% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_get.R
\name{load_get}
\alias{load_get}
\title{This wraps a GET request to the API}
\usage{
load_get(securityToken = NULL, documentType = NULL, processType = NULL,
  businessType = NULL, psrType = NULL, type_MarketAgreement.Type = NULL,
  contract_MarketAgreement.Type = NULL, auction.Type = NULL,
  auction.Category = NULL,
  classificationSequence_AttributeInstanceComponent.Position = NULL,
  outBiddingZone_Domain = NULL, biddingZone_Domain = NULL,
  controlArea_Domain = NULL, in_Domain = NULL, out_Domain = NULL,
  acquiring_Domain = NULL, timeInterval = NULL, periodStart = NULL,
  periodEnd = NULL, return_all = FALSE)
}
\description{
This wraps a GET request to the API
}

