% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_get_ts.R
\name{load_get_ts}
\alias{load_get_ts}
\title{This wraps a GET request to the API}
\usage{
load_get_ts(documentType = NULL, processType = NULL, businessType = NULL,
  psrType = NULL, type_MarketAgreement.Type = NULL,
  contract_MarketAgreement.Type = NULL, auction.Type = NULL,
  auction.Category = NULL,
  classificationSequence_AttributeInstanceComponent.Position = NULL,
  outBiddingZone_Domain = NULL, biddingZone_Domain = NULL,
  controlArea_Domain = NULL, in_Domain = NULL, out_Domain = NULL,
  acquiring_Domain = NULL, timeInterval = NULL, periodStart = NULL,
  periodEnd = NULL, securityToken = NULL)
}
\arguments{
\item{documentType}{The coded type of the document being sent. The 
document type identifies the principal characteristic of the 
status request. Refer to ENTSO-E 
Core Component Code list document for valid codes. The document 
type value must be exactly 3 alphanumeric characters (no blanks).
This information is mandatory. There's no dependency 
requirements.}
}
\description{
Function to download load data
}
