% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outages_get.R
\name{outages_get}
\alias{outages_get}
\title{This wraps a GET request to the API}
\usage{
outages_get(documentType = NULL, processType = NULL, businessType = NULL,
  psrType = NULL, type_MarketAgreement.Type = NULL,
  contract_MarketAgreement.Type = NULL, auction.Type = NULL,
  auction.Category = NULL,
  classificationSequence_AttributeInstanceComponent.Position = NULL,
  outBiddingZone_Domain = NULL, biddingZone_Domain = NULL,
  controlArea_Domain = NULL, in_Domain = NULL, out_Domain = NULL,
  acquiring_Domain = NULL, timeInterval = NULL, periodStart = NULL,
  periodEnd = NULL, securityToken = Sys.getenv("ENTSOE_PAT"))
}
\description{
There's a limit of 200 documents, so the API will return an
error if this limit is reached.
}
\examples{

library(entsoeR)

# Actual Total outages
# https://transparency.entsoe.eu/content/static_content/Static\%20content/web\%20api/Guide.html#_outages
# 4.6.1. Unavailability of Consumption Units [7.1A&B]
outages_get(documentType = "A76", 
            periodStart = "201702012300", 
            periodEnd = "201702172300", 
            BiddingZone_Domain = "10YCZ-CEPS-----N")
 
 # 4.6.2. Unavailability of Transmission Infrastructure [10.1.A&B]
 outages_get(documentType = "A78", 
             businessType = "A53",
             in_Domain = "10YCZ-CEPS-----N",
             out_Domain = "10YSK-SEPS-----K",
             periodStart = "201702162300", 
             periodEnd = "201702192300")
          
 # 4.6.3. Unavailability of Offshore Grid Infrastructure [10.1.C]
 outages_get(documentType = "A79", 
          biddingZone_Domain = "10YDE-EON------1",
          periodStart = "201512312300", 
          periodEnd = "201612312300")
 
 # 4.6.4. Unavailability of Generation Units [15.1.A&B]
 outages_get(documentType = "A80",
             businessType = "A53",
             biddingZone_Domain = "10YCZ-CEPS-----N",
             periodStart = "201512312300", 
             periodEnd = "201612312300")
 
 # French nukes
 outages_get(documentType = "A80",
             businessType = "A53",
             biddingZone_Domain = "10YFR-RTE------C",
             periodStart = "201711072300", 
             periodEnd = "201712312300")
 
 4.6.5. Unavailability of Production Units [15.1.C&D]
 outages_get(documentType = "A77",
             businessType = "A53",
             biddingZone_Domain = "10YCZ-CEPS-----N",
             periodStart = "201512312300", 
             periodEnd = "201612312300")

}
