% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_areas.R
\name{print_areas}
\alias{print_areas}
\title{Print Available Area Codes and Names}
\usage{
print_areas()
}
\value{
This function does not return a value but prints a table of area codes and area names to the console.
}
\description{
This function loads the \code{AreaCodes} dataset and prints a table of unique area codes and their corresponding names.
The output is intended to help users identify valid values for the \code{areas} argument in the \code{fetch} function.
}
\details{
The function loads the \code{AreaCodes} dataset from the \code{data/AreaCodes.Rda} file and
displays a unique list of area codes and names using the \code{knitr::kable()} function for formatted output.
}
\examples{
\dontrun{
print_areas()
}
}
