% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_technologies.R
\name{print_technologies}
\alias{print_technologies}
\title{Print Available Technologies}
\usage{
print_technologies()
}
\value{
prints a table of available technologies to the console.
}
\description{
This function prints a list of technologies that can be used as inputs for the
\code{technology} argument of the \code{fetch} function. It provides a reference for users to know
which technologies are available, and it also notes that the \code{technology} argument
can be left blank to include all technologies.
}
\details{
Outputs a list of all possible technologies.
These technologies can be used as valid inputs for the \code{technology} argument in the \code{fetch} function.
}
\examples{
\dontrun{
print_technologies()
}
}
