% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reset_folder_path.R
\name{reset_folder_path}
\alias{reset_folder_path}
\title{Reset the Stored Folder Path}
\usage{
reset_folder_path()
}
\description{
This function resets the stored folder path, clearing it from the current R session.
The next time the \code{fetch()} function is run, the user will be prompted to enter a new folder path.
}
