% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission_get.R
\name{transmission_get}
\alias{transmission_get}
\title{This wraps a GET request to the API}
\usage{
transmission_get(securityToken = NULL, documentType = NULL,
  processType = NULL, businessType = NULL, psrType = NULL,
  type_MarketAgreement.Type = NULL, contract_MarketAgreement.Type = NULL,
  auction.Type = NULL, auction.Category = NULL,
  classificationSequence_AttributeInstanceComponent.Position = NULL,
  outBiddingZone_Domain = NULL, biddingZone_Domain = NULL,
  controlArea_Domain = NULL, in_Domain = NULL, out_Domain = NULL,
  acquiring_Domain = NULL, timeInterval = NULL, periodStart = NULL,
  periodEnd = NULL, return_all = FALSE)
}
\description{
This wraps a GET request to the API
}
\examples{
transmission_get(securityToken = securityToken, documentType = "A11",
                 in_Domain = "10YCZ-CEPS-----N",
                 out_Domain = "10YSK-SEPS-----K",
                 periodStart = "201611232300",
                 periodEnd = "201612312300")

transmission_get(securityToken = securityToken, documentType = "A09",
                 in_Domain = "10YCZ-CEPS-----N",
                 out_Domain = "10YSK-SEPS-----K",
                 periodStart = "201611232300",
                 periodEnd = "201612312300")
}
