% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission_parse_timeseries_individual.R
\name{transmission_parse_timeseries_individual}
\alias{transmission_parse_timeseries_individual}
\title{Helper function to parse a single time series node.}
\usage{
transmission_parse_timeseries_individual(ts_individual)
}
\arguments{
\item{ts_individual}{A single time series node.}
}
\description{
The time series structure can be individual so I need to add the
possible nodes that can pop-up.
}

