% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsget.R
\name{tsget}
\alias{tsget}
\title{Splits request in one-year chunks for automated download}
\usage{
tsget(type, dateseq, documentType, processType, psrType, in_Domain,
  outBiddingZone_Domain = "10YCZ-CEPS-----N", securityToken)
}
\arguments{
\item{type}{data type (generation, load)}

\item{documentType}{The coded type of the document being sent. The 
document type identifies the principal characteristic of the 
status request. Refer to ENTSO-E 
Core Component Code list document for valid codes. The document 
type value must be exactly 3 alphanumeric characters (no blanks).
This information is mandatory. There's no dependency 
requirements.}
}
\description{
Wrapper function for load_get_ts and generation_get_ts
}
