% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_get.R
\name{load_get}
\alias{load_get}
\title{This wraps a GET request to the API}
\usage{
load_get(documentType = NULL, processType = NULL, businessType = NULL,
  psrType = NULL, type_MarketAgreement.Type = NULL,
  contract_MarketAgreement.Type = NULL, auction.Type = NULL,
  auction.Category = NULL,
  classificationSequence_AttributeInstanceComponent.Position = NULL,
  outBiddingZone_Domain = NULL, biddingZone_Domain = NULL,
  controlArea_Domain = NULL, in_Domain = NULL, out_Domain = NULL,
  acquiring_Domain = NULL, timeInterval = NULL, periodStart = NULL,
  periodEnd = NULL, securityToken = Sys.getenv("ENTSOE_PAT"))
}
\description{
This wraps a GET request to the API
}
\examples{

library(entsoeR)

# Actual Total Load
# https://transparency.entsoe.eu/content/static_content/Static\%20content/web\%20api/Guide.html#_load_domain
# 4.1.1
load_get(documentType = "A65", 
         processType = "A16", 
         periodStart = "201702012300", 
         periodEnd = "201702172300", 
         outBiddingZone_Domain = "10YCZ-CEPS-----N")
 
 # 4.1.2. Day-Ahead Total Load Forecast [6.1.B]
 load_get(documentType = "A65", 
          processType = "A01", 
          periodStart = "201702162300", 
          periodEnd = "201702192300", 
          outBiddingZone_Domain = "10YCZ-CEPS-----N")
          
 # 4.1.3. Week-Ahead Total Load Forecast [6.1.C]
 # This returns a wierd result, see the second
 # call where there are two different business types.
 load_get(documentType = "A65", 
          processType = "A32", 
          periodStart = "201702162300", 
          periodEnd = "201703192300", 
          outBiddingZone_Domain = "10YCZ-CEPS-----N")
 load_get(documentType = "A65", 
          processType = "A32", 
          periodStart = "201512312300", 
          periodEnd = "201612312300", 
          outBiddingZone_Domain = "10YCZ-CEPS-----N", return_all = TRUE)
 
 # 4.1.5. Year-Ahead Total Load Forecast [6.1.E]
 # NOT WORKING resolution not implemented.
 load_get(documentType = "A65", 
          processType = "A33", 
          periodStart = "201612312300", 
          periodEnd = "201712312300", 
          outBiddingZone_Domain = "10YCZ-CEPS-----N")

}
